[style] css 고정 좌표에서 center로 변경
