+++ src/shared/components/appointmentmanagementbox/AppointmentManagementBox.css.ts
@@ -2,31 +2,6 @@ import { style } from '@vanilla-extract/css';
 import { vars } from '@shared/styles/theme.css';
 import { recipe } from '@vanilla-extract/recipes';
 
-export const minibutton = recipe({
-  base: {
-    position: 'absolute',
-    right: '2.2rem',
-    bottom: '2.2rem',
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    gap: '1rem',
-
-    width: '4.6rem',
-    height: '2.4rem',
-    padding: '0.2rem 1.1rem',
-    flexShrink: 0,
-    borderRadius: '10rem',
-  },
-  variants: {
-    background: {
-      suggestedPromises: { background: vars.color.baroBlue },
-      votingPromises: { background: vars.color.yellow1 },
-      confirmedPromises: { background: vars.color.red1 },
-    },
-  },
-});
-
 export const container = style({
   display: 'flex',
   flexDirection: 'column',
@@ -61,7 +36,6 @@ export const topBox = recipe({
 
 export const bottomBox = style({
   display: 'flex',
-  position: 'relative',
 
   width: '100%',
   height: '13.5rem',
